CORE-538 Cannot create a DatabaseChangeLog Table in a Sybase database Standardize on "INT" for ORDEREXECUED
git-svn-id: http://liquibase.jira.com/svn/CORE/trunk@1266 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ liquibase-core/src/main/java/liquibase/database/AbstractDatabase.java
@@ -352,9 +352,9 @@ public abstract class AbstractDatabase implements Database {
             }
             if (!hasOrderExecuted) {
                 executor.comment("Adding missing databasechangelog.orderexecuted column");
-                statementsToExecute.add(new AddColumnStatement(getLiquibaseSchemaName(), getDatabaseChangeLogTableName(), "ORDEREXECUTED", "BIGINT", null, new UniqueConstraint()));
+                statementsToExecute.add(new AddColumnStatement(getLiquibaseSchemaName(), getDatabaseChangeLogTableName(), "ORDEREXECUTED", "INT", null, new UniqueConstraint()));
                 statementsToExecute.add(new UpdateStatement(getLiquibaseSchemaName(), getDatabaseChangeLogTableName()).addNewColumnValue("ORDEREXECUTED", -1));
-                statementsToExecute.add(new SetNullableStatement(getLiquibaseSchemaName(),  getDatabaseChangeLogTableName(), "ORDEREXECUTED", "BIGINT", false));
+                statementsToExecute.add(new SetNullableStatement(getLiquibaseSchemaName(),  getDatabaseChangeLogTableName(), "ORDEREXECUTED", "INT", false));
             }
             if (checksumNotRightSize) {
                 executor.comment("Modifying size of databasechangelog.md5sum column");
